Сompleted rearrangement of components.
