+++ src/components/plan-bills/Planing.js
@@ -3,7 +3,7 @@ import AddBill from './AddBill';
 import BillList from './BillList';
 import BillTotal from './BillTotal';
 import BillOptions from './BillOptions';
-import { BillContext } from '../context/BillContext';
+import { BillContext } from '../../context/BillContext';
 import EditBills from './EditBills';
 
 function Planing() {
