+++ src/lang/ua.json
@@ -1,0 +1,42 @@
+{
+    "greeting": "Привіт, {name}",
+    "myMessage": "Шось українською {date, time, ::MMdd}",
+    "addIncomeHere": "Внесіть свої витрати або доходи сюди:",
+    "designedBy": "Розроблено Людмилою Шуландт. Права захищені.",
+    "home": "Головна",
+    "planing": "Планування",
+    "statistics": "Статистика",
+    "categories": "Категорії",
+    "day": "День",
+    "month": "Місяць",
+    "planYourBills": "Запланувати витрати",
+    "editCategories": "Редагувати категорії",
+    "dayStatistic": "Статистика за дату",
+    "monthStatistic": "Статистика за місяць",
+    "yearStatistic": "Статистика за рік",
+    "catStatistic": "Щомісячна статистика по категоріям",
+    "pageIsNotFound": "Сторінку не знайдено",
+    "year": "Рік",
+    "chooseCategory": "Виберіть категорію",
+    "plan": "Запланувати",
+    "daily": "За день",
+    "monthly": "За місяць",
+    "yearly": "За рік",
+    "billCost": "усього нараховано ",
+    "saved": " економії ",
+    "total": "Усього: ",
+    "edit": "Редагувати",
+    "done": "Готово",
+    "delete": "ВИДАЛИТИ",
+    "add": "ДОДАТИ",
+    "addDeleteCategory": "Редагуй список категорій",
+    "registration": "Регістрація",
+    "registrationBy": "Зареєструйся через Фейсбук або за допомогою Гугл-пошти: ",
+    "category": "Категорія",
+    "title": "Примітки",
+    "price": "Ціна",
+    "date": "Дата",
+    "areU": "Ти домогосподарка і шукаєш просте рішення для підрахунку своїх витрат?",
+    "see": "Дивись як просто та зрозуміло можна керувати своїми фінансами у нашому застосунку:"
+}
+
